Add priority display
