+++ services/frontend/src/components/shared/controls/priority-picker.js
@@ -1,15 +1,8 @@
 import _ from 'lodash';
 import React from 'react';
 import PropTypes from 'prop-types';
-import { FormControl, FormLabel, FormControlLabel, RadioGroup, Radio, colors } from '@material-ui/core';
-
-const priorities = {
-    'Very high': [2, colors.deepOrange[500]],
-    High: [1, colors.orange[500]],
-    Normal: [0, colors.yellow[500]],
-    Low: [-1, colors.green[500]],
-    'Very low': [-2, colors.blue[500]],
-};
+import { FormControl, FormLabel, FormControlLabel, RadioGroup, Radio } from '@material-ui/core';
+import priorities from '../priority-constants';
 
 class PriorityPicker extends React.Component {
     constructor(props) {
