+++ services/frontend/src/components/shared/index.js
@@ -5,3 +5,4 @@ export { default as Page } from './page';
 export { default as Navbar } from './navbar';
 export { default as Sidebar } from './sidebar';
 export { default as CourseLabel } from './course-label';
+export { default as PriorityDisplay } from './priority-display';
